+++ sbpy/spectroscopy/tests/test_sources.py
@@ -5,17 +5,15 @@ import numpy as np
 import astropy.units as u
 from astropy.modeling.models import BlackBody
 import astropy.constants as const
-import synphot
-from .. import sources
-from ..sources import (BlackbodySource, SpectralSource, SynphotRequired,
-                       Reddening)
+synphot = pytest.importorskip("synphot")
+
+from ..sources import (BlackbodySource, SpectralSource, Reddening)
 from ..core import SpectralGradient
 from ...photometry import bandpass
 from ...units import hundred_nm
 
 V = bandpass('johnson v')
 I = bandpass('cousins i')
-
 
 class Star(SpectralSource):
     def __init__(self):
@@ -24,11 +22,6 @@ class Star(SpectralSource):
 
 
 class TestSpectralSource:
-    def test_init_error(self, monkeypatch):
-        with pytest.raises(SynphotRequired):
-            monkeypatch.setattr(sources, 'synphot', None)
-            Star()
-
     @pytest.mark.parametrize('wfb, interpolate', (
         ([V], False),
         ([1] * u.um, True),